move gesture to package
