+++ photoview/src/main/java/com/github/chrisbanes/photoview/gesture/CustomGestureDetector.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package com.github.chrisbanes.photoview;
+package com.github.chrisbanes.photoview.gesture;
 
 import android.content.Context;
 import android.view.MotionEvent;
@@ -21,12 +21,13 @@ import android.view.ScaleGestureDetector;
 import android.view.VelocityTracker;
 import android.view.ViewConfiguration;
 
+import com.github.chrisbanes.photoview.Util;
 import com.github.chrisbanes.photoview.listener.OnGestureListener;
 
 /**
  * Does a whole lot of gesture detecting.
  */
-class CustomGestureDetector {
+public class CustomGestureDetector {
 
     private static final int INVALID_POINTER_ID = -1;
 
@@ -42,7 +43,7 @@ class CustomGestureDetector {
     private final float mMinimumVelocity;
     private OnGestureListener mListener;
 
-    CustomGestureDetector(Context context, OnGestureListener listener) {
+    public CustomGestureDetector(Context context, OnGestureListener listener) {
         final ViewConfiguration configuration = ViewConfiguration
                 .get(context);
         mMinimumVelocity = configuration.getScaledMinimumFlingVelocity();
